Lesson 15
Task 1
Create a collection of integers and fill it with values entered from the console.
When populating a collection with numbers from the console, you should make type casts.
The code to convert a string to a number can be used as follows:
int i = Integer.parseInt(str);
or
int val = Integer.valueOf(str);
To end your input, enter the word "exit".
When converting a string to a number, you should consider the possibility of exceptions.
Some useful code examples for casting a string to a number can be found here
https://www.geeksforgeeks.org/how-to-convert-a-string-to-an-int-in-java/
Display even numbers from the collection on the screen.

Task 2
Create a collection of shapes.
(The classes for the shapes can be taken from previous homework assignments. It is not necessary to use sealed classes)
Fill the collection with various shapes.
Loop through the collection and call a method to calculate and display the perimeter of each shape.

Task 3
Create a collection of integers.
Fill the collection with random numbers.
Let the collection size be set from the console.
Expect possible errors when retrieving the collection size from the console.
Provide a check for validity of the entered collection size.
Calculate and display the arithmetic mean of all elements in the collection.

Task 6
Create a collection of unique names of all students in our group + teacher.
Display the collection on the screen.